refactor(analyze): extract _plan_output_files → output_plan.py (CH-07)
Pure internal reorganization (move + re-export; zero behavior change), part 2 of the
oversized-module split. analyze.py 1507 → 1276.

- agentteams/output_plan.py — _plan_output_files (the agents/instructions/references/
  builder output-file planner). The three analyze-owned symbols it uses
  (GOVERNANCE_AGENTS, ALWAYS_INCLUDED_DOMAIN_AGENTS, _dedupe_keep_order) are reused
  across analyze, so they stay there and output_plan imports them LAZILY inside the
  function — keeping the graph acyclic (analyze imports _plan_output_files at module top).
- analyze.py re-exports _plan_output_files so analyze._plan_output_files (and cli/generate's
  namespace use) resolve unchanged.

LENGTH_ALLOWLIST comment refreshed (1507 → 1276); output_plan.py added to _REFACTOR_MODULES
(CH-22 annotation ratchet). No CHANGELOG (pure reorg). Full suite 1385 passed; hygiene + RSR1
green. Adversarial+conflict audited together with the emit split (scratch extraction, zero
regressions).
